backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
modal>=0.55.0
python-dotenv>=1.0.0
pydantic>=2.5.0
aiofiles>=23.2.0
httpx>=0.25.0 